Snapshot worker
