+++ bee-message/Cargo.toml
@@ -21,7 +21,6 @@ bytemuck = "1.2"
 digest = "0.9"
 hex = "0.4"
 serde = "1.0"
-thiserror = "1.0"
 
 [dev-dependencies]
 bee-test = { path = "../bee-test" }
