+++ bee-node/src/node/builder.rs
@@ -56,11 +56,7 @@ fn ctrl_c_listener() -> oneshot::Receiver<()> {
 }
 
 #[derive(Error, Debug)]
-pub enum Error {
-    /// Occurs, when there is an error while reading the snapshot file.
-    #[error("Reading snapshot file failed: {0}")]
-    SnapshotError(bee_snapshot::Error),
-}
+pub enum Error {}
 
 pub struct BeeNodeBuilder<B: Backend> {
     deps: HashMap<TypeId, &'static [TypeId]>,
@@ -174,21 +170,13 @@ impl<B: Backend> NodeBuilder<BeeNode<B>> for BeeNodeBuilder<B> {
         let this = this.with_resource(ShutdownStream::new(ctrl_c_listener(), events));
 
         info!("Initializing snapshot handler...");
-        let (this, snapshot) = bee_snapshot::init::<BeeNode<B>>(&config.snapshot, network_id, this)
-            .await
-            .map_err(Error::SnapshotError)?;
+        let this = bee_snapshot::init::<BeeNode<B>>(&config.snapshot, network_id, this).await;
 
         // info!("Initializing ledger...");
         // let mut this = bee_ledger::init::<BeeNode<B>>(snapshot.header().ledger_index(), this);
 
         info!("Initializing protocol layer...");
-        let this = init::<BeeNode<B>>(
-            config.protocol.clone(),
-            config.storage.clone(),
-            snapshot,
-            network_id,
-            this,
-        );
+        let this = init::<BeeNode<B>>(config.protocol.clone(), config.storage.clone(), network_id, this);
 
         let mut this = this.with_worker::<VersionCheckerWorker>();
         this = this.with_worker_cfg::<Mqtt>(config.mqtt);
