+++ bee-snapshot/Cargo.toml
@@ -24,4 +24,5 @@ futures = "0.3"
 log = "0.4"
 reqwest = { version = "0.10", features = ["stream"] }
 serde = { version = "1.0", features = ["derive" ] }
+thiserror = "1.0"
 tokio = "0.2"
